move `getCurrentCurrency` from `ui/selectors/selectors.js` to `ui/ducks/metamask/metamask.js`
I moved `getCurrentCurrency` into `ui/ducks/` instead of kept in `ui/selectors/` because `ui/ducks/metamask/metamask.js` already exports a `getNativeCurrency`, so it seems to fit in well there.
