+++ ui/components/multichain/asset-picker-amount/asset-picker-modal/asset-picker-modal.tsx
@@ -32,7 +32,6 @@ import { AssetType } from '../../../../../shared/constants/transaction';
 import {
   getAllTokens,
   getCurrentChainId,
-  getCurrentCurrency,
   getNativeCurrencyImage,
   getSelectedAccountCachedBalance,
   getSelectedInternalAccount,
@@ -42,6 +41,7 @@ import {
 } from '../../../../selectors';
 import {
   getConversionRate,
+  getCurrentCurrency,
   getNativeCurrency,
 } from '../../../../ducks/metamask/metamask';
 import { useTokenTracker } from '../../../../hooks/useTokenTracker';
