+++ ui/ducks/metamask/metamask.js
@@ -580,3 +580,7 @@ export function doesUserHaveALedgerAccount(state) {
     return kr.type === KeyringType.ledger;
   });
 }
+
+export function getCurrentCurrency(state) {
+  return state.metamask.currentCurrency;
+}
