+++ ui/pages/asset/components/asset-page.tsx
@@ -6,7 +6,6 @@ import { EthMethod } from '@metamask/keyring-api';
 import { isEqual } from 'lodash';
 import { Hex } from '@metamask/utils';
 import {
-  getCurrentCurrency,
   getIsBridgeChain,
   getIsSwapsChain,
   getSelectedInternalAccount,
@@ -39,7 +38,10 @@ import TokenCell from '../../../components/app/assets/token-cell';
 import TransactionList from '../../../components/app/transaction-list';
 import { getPricePrecision, localizeLargeNumber } from '../util';
 import { DEFAULT_ROUTE } from '../../../helpers/constants/routes';
-import { getConversionRate } from '../../../ducks/metamask/metamask';
+import {
+  getConversionRate,
+  getCurrentCurrency,
+} from '../../../ducks/metamask/metamask';
 import { toChecksumHexAddress } from '../../../../shared/modules/hexstring-utils';
 import CoinButtons from '../../../components/app/wallet-overview/coin-buttons';
 import { getIsNativeTokenBuyable } from '../../../ducks/ramps';
