+++ ui/pages/swaps/prepare-swap-page/review-quote.js
@@ -52,7 +52,6 @@ import {
 import {
   conversionRateSelector,
   getSelectedAccount,
-  getCurrentCurrency,
   getTokenExchangeRates,
   getSwapsDefaultToken,
   getCurrentChainId,
@@ -65,7 +64,11 @@ import {
   getSmartTransactionsOptInStatus,
   getSmartTransactionsEnabled,
 } from '../../../../shared/modules/selectors';
-import { getNativeCurrency, getTokens } from '../../../ducks/metamask/metamask';
+import {
+  getNativeCurrency,
+  getTokens,
+  getCurrentCurrency,
+} from '../../../ducks/metamask/metamask';
 import {
   setCustomApproveTxData,
   showModal,
