+++ ui/selectors/selectors.js
@@ -706,10 +706,6 @@ export function getNftDetectionEnablementToast(state) {
   return state.appState.showNftDetectionEnablementToast;
 }
 
-export function getCurrentCurrency(state) {
-  return state.metamask.currentCurrency;
-}
-
 export function getTotalUnapprovedCount(state) {
   return state.metamask.pendingApprovalCount ?? 0;
 }
